Add xml encoding to text
